server/requirements.txt
flask==3.1.2
Flask-Cors==6.0.1
python-docx==1.2.0
requests==2.32.5
psycopg2-binary==2.9.10
mammoth==1.10.0
nltk==3.9.1
sentence-transformers==5.1.0
python-dotenv==1.1.1
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
flask-migrate==4.0.7
flask-sqlalchemy==3.1.1
celery==5.3.6
redis==5.0.4
apscheduler==3.10.4
twilio==9.2.3
sendgrid==6.11.0
googletrans==4.0.0rc1
langdetect==1.0.9
pytz==2024.1
croniter==2.0.1
pydantic==2.8.2
marshmallow==3.22.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
gunicorn==22.0.0
